requirements.txt
Flask==0.11.1
flask-restplus==0.9.2
protobuf==3.0.0
grpcio==1.6.0
google-cloud==0.27.0
google-cloud-happybase==0.26.0
google-api-python-client==1.6.3
gunicorn==19.6.0
python-dateutil==2.5.3
python-dotenv==0.6.0
universal-analytics-python==0.2.4
